Adding landsfcutil library to the party.
This was a pretty standard addition: no real tricks to it. Instead of calling the script
make_landsfcutil_lib.sh we go to the src/ directory and make it directly. Made some modifications
to the Makefile to allow this. Deleted the relics of the old build system to avoid problems.

This only works for cheyenne, and has only been tested with intel 18.0.1 for now. Should not be
hard to adapt for other platforms however.