nemsiogfs source code changes required for gfortran compilation
The nemsiogfs source code contains non-standard fortran that gfortran
does not accept. aero_tracername is a character array with non-uniform
element sizes, which is not allowed according to any fortran standard.

The fix is fairly simple: pad each element with spaces to make them
all the same length, then whenever each element is called later call it
with a TRIM() command. This should produce identical results to what
the developers probably intended.