Rearrange how '-a' flag works for make script:
Back when I first introduced this flag I had no idea how these libraries worked or which libraries were needed for what applications. Now I have a much better idea: the initial set of libraries were for the FV3 global, the ones I added later were for FV3 stand-alone regional (SAR).

Now it is time to add libraries needed for UPP, and I will take this opportunity to re-vamp how the -a flag works. Instead of the old, clunky compile "some" or "all" (with 0 or 1, 0 being default), users can specify the application they need the libraries for. Options are global (the old default, -a 0), sar (the old -a 1), upp (a new implementation, still incomplete but will be done at the end of this larger effort), and all (all libraries)

I kept in the ability to use the old flags (0 or 1) for back-compatability but these are no longer documented.

I attempted to check if the new and old libraries were bit-for-bit identical, but this does not appear to be possible
since the compiled libraries differ even with the same source code; perhaps related to directory names?