Rearrange how '-a' flag works for make script (#8)
Improving how the -a flag works. Instead of the old, clunky compile "some" or "all" (with 0 or 1, 0 being default), users can specify the application they need the libraries for. Options are global (the old default, -a 0), sar (the old -a 1), upp (a new implementation, still incomplete but will be done at the end of this larger effort), and all (all libraries)

Kept in the ability to use the old flags (0 or 1) for back-compatability but these are no longer documented.