Bug fix for building libsp without openmp:
When you do not build with openmp flags (at least for gnu, haven't tested other compilers), 'OMP_GET_THREAD_NUM' and 'OMP_GET_NUM_THREADS' can not be called because these are openmp function calls, and if we are not building with openmp flags these may not be callable.

This commit includes a flag and a CPP instruction so that when building without openmp, these functions do not get called, and the calling routine which expects the number of threads always gets 1 as an answer.